tasks offloaded into BG queue with default priority

+++ app/Http/Controllers/TopicCategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\TopicCategory;
 use App\Http\Resources\Topics\TopicCategory as TopicCategoryResources;
+use App\Jobs\ProcessTopicCategory;
 use Illuminate\Http\Request;
 
 class TopicCategoryController extends Controller
@@ -18,17 +19,28 @@ class TopicCategoryController extends Controller
         return TopicCategoryResources::collection(TopicCategory::all());
     }
 
-    public function store(Request $request)
+    public static function store($request)
     {
         $category = new TopicCategory;
-        $category->label = $request->label;
+        $category->label = $request['label'];
         $category->save();
+    }
+
+    public static function delete($request) {
+        $categoryToDelete = TopicCategory::find($request['category_id']);
+        if ($categoryToDelete)
+            $categoryToDelete->delete();
+    }
+
+    public function dispatchStore(Request $request)
+    {
+        ProcessTopicCategory::dispatch($request->all(), 'store')->onQueue('default');
         return response()->json(['status' => 'success'], 200);
     }
 
-    public function delete(Request $request) {
-        $categoryToDelete = TopicCategory::find($request->category_id);
-        $categoryToDelete->delete();
-        return;
+    public function dispatchDelete(Request $request)
+    {
+        ProcessTopicCategory::dispatch($request->all(), 'delete')->onQueue('default');
+        return response()->json(['status' => 'success'], 200);
     }
 }
