+++ app/Http/Controllers/TopicController.php
@@ -6,6 +6,7 @@ use App\Topic;
 use App\Traits\ForumTraits;
 use App\Http\Resources\Topics\Topic as TopicResources;
 use Illuminate\Http\Request;
+use App\Jobs\ProcessTopic;
 
 class TopicController extends Controller
 {
@@ -38,20 +39,31 @@ class TopicController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public static function store($request)
     {
         $topic = new Topic;
-        $topic->user_id = $request->user_id;
-        $topic->category_id = $request->category_id;
-        $topic->title = $request->title;
-        $topic->body = $request->body;
+        $topic->user_id = $request['user_id'];
+        $topic->category_id = $request['category_id'];
+        $topic->title = $request['title'];
+        $topic->body = $request['body'];
         $topic->save();
-        return;
     }
 
-    public function delete(Request $request) {
-        $topicToDelete = Topic::find($request->topic);
-        $topicToDelete->delete();
-        return;
+    public static function delete($request) {
+        $topicToDelete = Topic::find($request['topic']);
+        if ($topicToDelete)
+            $topicToDelete->delete();
+    }
+
+    public function dispatchStore(Request $request)
+    {
+        ProcessTopic::dispatch($request->all(), 'store')->onQueue('default');
+        return response()->json(['status' => 'success'], 200);
+    }
+
+    public function dispatchDelete(Request $request)
+    {
+        ProcessTopic::dispatch($request->all(), 'delete')->onQueue('default');
+        return response()->json(['status' => 'success'], 200);
     }
 }
